requirements.txt
Flask==2.3.2
pyodbc==5.1.0
python-dotenv==1.0.1
